+++ src/app/common/components/dynamic-table/dynamic-table.component.ts
@@ -8,10 +8,13 @@
 import { AfterContentInit, ChangeDetectionStrategy, Component, inject, Input, OnChanges, OnInit } from "@angular/core";
 import { MatTableDataSource, MatTableModule } from "@angular/material/table";
 import { DataRow, DataSchemaField, OperationColumnClassEnum } from "src/app/interface/dataset.interface";
-import { TableSourceRowInterface } from "./dynamic-table.interface";
+import { OdfExtraAttributes, TableSourceRowInterface } from "./dynamic-table.interface";
 import { NgFor, NgClass, NgIf } from "@angular/common";
 import { ClipboardModule } from "@angular/cdk/clipboard";
 import { ToastrModule, ToastrService } from "ngx-toastr";
+import { OdfTypeMapper } from "../../helpers/data.helpers";
+import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import { AttributesSchemaModalComponent } from "./components/attributes-schema-modal/attributes-schema-modal.component";
 
 @Component({
     selector: "app-dynamic-table",
@@ -41,6 +44,7 @@ export class DynamicTableComponent implements OnInit, OnChanges, AfterContentIni
     public displayedColumns: string[] = [];
     public readonly OperationColumnClassEnum: typeof OperationColumnClassEnum = OperationColumnClassEnum;
     private toastr = inject(ToastrService);
+    private ngbModalService = inject(NgbModal);
 
     public ngOnInit(): void {
         this.displayTable();
@@ -66,10 +70,26 @@ export class DynamicTableComponent implements OnInit, OnChanges, AfterContentIni
 
             // Special case: displaying schema itself
         } else if (!this.dataRows) {
-            this.dataSource.data = this.schemaFields;
-            const arrFieldsLength = this.schemaFields.map((item) => Object.keys(item).length);
-            const indexFieldMaxLength = arrFieldsLength.indexOf(Math.max.apply(null, arrFieldsLength));
-            this.displayedColumns = Object.keys(this.schemaFields[indexFieldMaxLength]);
+            this.dataSource.data = this.schemaFields.map((x) => {
+                return {
+                    name: { value: x.name, cssClass: "" },
+                    type: {
+                        value:
+                            x.extra && OdfExtraAttributes.EXTRA_ATTRIBUTE_TYPE in x.extra
+                                ? x.extra[OdfExtraAttributes.EXTRA_ATTRIBUTE_TYPE].kind
+                                : OdfTypeMapper(x.type),
+                        cssClass: "",
+                    },
+                    description: {
+                        value:
+                            x.extra && OdfExtraAttributes.EXTRA_ATTRIBUTE_DESCRIPTION in x.extra
+                                ? x.extra[OdfExtraAttributes.EXTRA_ATTRIBUTE_DESCRIPTION]
+                                : "",
+                        cssClass: "",
+                    },
+                };
+            });
+            this.displayedColumns = ["name", "type", "description"];
 
             // Casual case, displaying data
         } else {
@@ -80,5 +100,23 @@ export class DynamicTableComponent implements OnInit, OnChanges, AfterContentIni
 
     public showCopyMessage() {
         this.toastr.success(`Copied`);
+    }
+
+    public showBadge(indexRow: number, indexColumn: number, data: DataSchemaField[]): boolean {
+        return (
+            indexColumn === 1 &&
+            !this.hasData &&
+            (Boolean(Object.keys(data[indexRow].extra ?? {}).length) || Boolean(data[indexRow].type?.fields?.length))
+        );
+    }
+
+    public showModal(index: number, data: DataSchemaField[]): void {
+        const modalRef = this.ngbModalService.open(AttributesSchemaModalComponent, {
+            size: "lg",
+            centered: true,
+            scrollable: true,
+        });
+        const modalRefInstance = modalRef.componentInstance as AttributesSchemaModalComponent;
+        modalRefInstance.element = data[index];
     }
 }