+++ CHANGELOG.md
@@ -8,10 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Lazy loading of cards when scrolling for account filters.
 - "Run Selection" context menu action in SQL query editor
+- Support for displaying column annotations
 ### Fixed
 - CORS error due to incorrect configuration when executing the SQL query
 - Corrected redirection to a page from a link in an email when the token has expired 
 - Activity tab: search functionality when switching process statuses
+### Changed
+- The dataset schema has been converted to ODF format 
 - Flows tab: the status filter was incorrectly initialized for request